Refactor before memoing
