+++ components/lib/ThreeComponent.tsx
@@ -8,6 +8,7 @@ import {TrackballControls} from "three/examples/jsm/controls/TrackballControls";
 import {ThreeControlType} from "@/src/types/general";
 import {Grass} from "@/src/utils/grass/grass";
 import {renderEnvironment} from "@/src/utils/background";
+import {getArrowHelper, getGroundPlane, getMeshForItem} from "@/src/utils/model";
 
 
 // TODO: Move this all to React.useMemo to preserve data over re-renders
@@ -97,48 +98,6 @@ export default function ThreeComponent({
                 controls.maxPolarAngle = Math.PI / 2;
         }
     }
-
-    const getMeshForItem = (item: AssetObject): THREE.Mesh => {
-        let model;
-        let material = new THREE.MeshBasicMaterial({ color: item.color ?
-                new THREE.Color(item.color) : 0x000000 });
-        let geometry;
-        let position1, position2;
-        switch (item.type) {
-            case "rect":
-                const rect = item as Rectangle;
-                geometry = new THREE.BoxGeometry(rect.w, rect.h, Math.round((rect.w + rect.h) / 2));
-                break;
-            case "circle":
-                geometry = new THREE.SphereGeometry((item as Circle).radius, 32, 16 );
-                break;
-            case "line":
-                const line = item as Line;
-                position1 = new THREE.Vector3(line.x1, line.y1, 0);
-                position2 = new THREE.Vector3(line.x2, line.y2, 0);
-                const height = position1.distanceTo(position2);
-
-                geometry = new THREE.CylinderGeometry( 5, 5, height, 32 );
-        }
-        model = new THREE.Mesh(geometry, material);
-        if (model && position1 && position2) {
-            const positionMid = new THREE.Vector3();
-            positionMid.addVectors(position1, position2).multiplyScalar(0.5);
-            model.position.copy(positionMid);
-            const direction = new THREE.Vector3();
-            direction.subVectors(position2, position1).normalize();
-
-            const quaternion = new THREE.Quaternion();
-            quaternion.setFromUnitVectors(new THREE.Vector3(0, 1, 0), direction);
-            model.setRotationFromQuaternion(quaternion);
-        } else if (model && item.type === "rect") {
-            const rect = item as Rectangle;
-            model.position.set(rect.x + rect.w / 2, rect.y + rect.h / 2, rect.z || 0);
-        } else if (model && typeof item.x === 'number' && typeof item.y === "number") {
-            model.position.set(item.x, item.y, item.z || 0);
-        }
-        return model;
-    };
 
     const refreshSceneItems = () => {
         if (scene) {
@@ -236,22 +195,7 @@ export default function ThreeComponent({
     };
 
     const addArrowHelper = ()=>{
-        const arrowGroup = new THREE.Group();
-        arrowGroup.name = "arrows";
-        const xAxisDirection = new THREE.Vector3(1, 0, 0);
-        const yAxisDirection = new THREE.Vector3(0, 1, 0);
-        const zAxisDirection = new THREE.Vector3(0, 0, 1);
-
-        const origin = new THREE.Vector3( 0, 0, 0 );
-        const length = 100;
-
-        const xAxisArrow = new THREE.ArrowHelper(xAxisDirection, origin, length, 0xff0000);
-        const yAxisArrow = new THREE.ArrowHelper(yAxisDirection, origin, length, 0x00ff00);
-        const zAxisArrow = new THREE.ArrowHelper(zAxisDirection, origin, length, 0x0000ff);
-
-        arrowGroup.add(xAxisArrow);
-        arrowGroup.add(yAxisArrow);
-        arrowGroup.add(zAxisArrow);
+        const arrowGroup = getArrowHelper();
 
         scene.add(arrowGroup);
 
@@ -368,30 +312,13 @@ export default function ThreeComponent({
 
     };
 
-    const addBasePlane = () => {
-        return new Promise(resolve => {
-            if (!scene || scene.children.find(mesh => mesh.name === "plane")) {
-                return resolve(false);
-            }
-            const geometry = new THREE.PlaneGeometry( planeSize, planeSize );
-            const material = new THREE.MeshBasicMaterial( {color: 0xffff00, side: THREE.DoubleSide} );
-            const loader = new THREE.TextureLoader();
-            loader.load(ground ,
-                function ( texture ) {
-                    texture.wrapS = texture.wrapT = THREE.RepeatWrapping;
-                    texture.offset.set( 0, 0 );
-                    texture.repeat.set( 2, 2 );
-                    material.map = texture;
-                    material.needsUpdate = true;
-                    const plane = new THREE.Mesh( geometry, material );
-                    //plane.position.setX(planeSize / 2);
-                    //plane.position.setY(planeSize / 2);
-                    plane.position.setZ(-1);
-                    plane.name = "plane";
-                    scene.add( plane );
-                    resolve(plane);
-                });
-        });
+    const addBasePlane = async () => {
+        if (!scene || scene.children.find(mesh => mesh.name === "plane")) {
+            return false;
+        }
+        const plane = await getGroundPlane(planeSize, planeSize, ground);
+        scene.add(plane);
+        return plane;
     };
 
     const setEnvironment = () => {
